third_party/WebKit/LayoutTests/shadow-dom/nodetree-radio-node-list.html

<!DOCTYPE html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
    var host = document.createElement('div');
    var root = host.attachShadow({mode: 'open'});
    root.innerHTML = '<form id="f1"><input name="n"><select name="n"></select></form>';
    assert_equals(root.firstChild.elements.n.length, 2);
    assert_equals(root.firstChild.n.length, 2);

    document.body.appendChild(host);
    assert_equals(root.firstChild.elements.n.length, 2);
    assert_equals(root.firstChild.n.length, 2);
}, 'RadioNodeList should work in Shadow DOM.');
</script>
